+++ packages/core/src/core/adapters/partial_fetch_targets.ts
